requirements.txt
Django>=5.1.4
psycopg2>=2.8
asgiref==3.8.1
djangorestframework==3.15.2
sqlparse==0.5.2
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg2==2.9.10
sqlparse==0.5.2